+++ packages/core/src/components/CardTable/Body/Body.tsx
@@ -1,4 +1,3 @@
-import { WithStyle } from '@medly-components/utils';
 import React, { FC, useCallback, useMemo } from 'react';
 import Text from '../../Text';
 import Cell from './Cell';
@@ -6,14 +5,13 @@ import Row from './Row';
 import { NoResult } from './Row/Row.styled';
 import { Props } from './types';
 
-
-export const Body: FC<Props> & WithStyle = React.memo(props => {
+export const Body: FC<Props> = React.memo(props => {
     const { data, columns, onRowClick, rowClickDisableKey, uniqueKeyName, withWhiteBackground } = props;
 
     const gridTemplateColumns = useMemo(() => columns.reduce((acc, curr) => acc.concat(`${curr.fraction || 1}fr `), ''), [columns]),
         handleRowClick = useCallback(
             (rowData: any) => {
-                return onRowClick && !rowData[rowClickDisableKey] ? () => onRowClick(rowData) : undefined;
+                return onRowClick && !rowData[rowClickDisableKey!] ? () => onRowClick(rowData) : undefined;
             },
             [onRowClick, rowClickDisableKey]
         );
@@ -21,7 +19,9 @@ export const Body: FC<Props> & WithStyle = React.memo(props => {
     if (data.length === 0) {
         return (
             <NoResult withWhiteBackground={withWhiteBackground} gridTemplateColumns={gridTemplateColumns}>
-                <Text textVariant="body2" textWeight="Strong" > No Result </Text>
+                <Text textVariant="body2" textWeight="Strong">
+                    No Result
+                </Text>
             </NoResult>
         );
     }
@@ -30,19 +30,19 @@ export const Body: FC<Props> & WithStyle = React.memo(props => {
         <tbody>
             {data.map((row, index) => (
                 <Row
-                    key={row[uniqueKeyName] || index}
+                    key={row[uniqueKeyName!] || index}
                     onClick={handleRowClick(row)}
                     withWhiteBackground={withWhiteBackground}
-                    isRowClickDisabled={row[rowClickDisableKey]}
+                    isRowClickDisabled={row[rowClickDisableKey!]}
                     gridTemplateColumns={gridTemplateColumns}
                 >
                     {columns.map(column => (
                         <Cell
                             key={column.field}
-                            isRowClickDisabled={row[rowClickDisableKey]}
+                            isRowClickDisabled={row[rowClickDisableKey!]}
                             data={row[column.field]}
                             column={column}
-                            rowId={row[uniqueKeyName]}
+                            rowId={row[uniqueKeyName!]}
                             rowData={row}
                         />
                     ))}