+++ packages/core/src/components/CheckboxGroup/CheckboxGroup.stories.tsx
@@ -3,14 +3,14 @@ import Button from '../Button';
 import { CheckboxGroup } from './CheckboxGroup';
 import { CheckboxGroupProps } from './types';
 
-export const sizes: CheckboxGroupProps['size'][] = ['XS', 'S', 'M', 'L', 'XL'];
+export const sizes: Required<CheckboxGroupProps>['size'][] = ['XS', 'S', 'M', 'L', 'XL'];
 
-export const FormWithCheckboxGroup: React.FC = (props: CheckboxGroupProps) => {
-    const [bikes, setBikes] = useState([]),
+export const FormWithCheckboxGroup: React.FC<CheckboxGroupProps> = props => {
+    const [bikes, setBikes] = useState<string[]>([]),
         handleSubmit = (e: React.FormEvent) => e.preventDefault();
     return (
         <form onSubmit={handleSubmit}>
-            <CheckboxGroup fullWidth required values={bikes} onChange={setBikes} {...props} />
+            <CheckboxGroup fullWidth required values={bikes} {...props} onChange={setBikes} />
             <Button size="S" type="submit">
                 Submit
             </Button>