+++ packages/core/src/components/Tabs/types.ts
@@ -31,7 +31,7 @@ export interface TabsProps extends HTMLProps<HTMLDivElement> {
 }
 
 export interface StaticProps {
-    Tab?: React.FC<TabProps> & WithStyle;
-    TabList?: React.FC<TabListProps> & WithStyle;
-    TabPanel?: React.FC<TabPanelProps> & WithStyle;
+    Tab: React.FC<TabProps> & WithStyle;
+    TabList: React.FC<TabListProps> & WithStyle;
+    TabPanel: React.FC<TabPanelProps> & WithStyle;
 }